test: move openapi:reduce tests over
